backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
geoalchemy2
geopandas
rasterio
shapely
scikit-learn
xgboost
pulp
scipy
python-multipart
bcrypt
python-jose[cryptography]
python-multipart
fastapi-mail
python-dotenv
